avm2: Move avm2 playerglobal generation into its own file, out of lib.rs
